chromium-integration
run nw as node

decouple node dll

add g_dom_context

new mac loop

fixup! chromium-integration

fixup! new mac loop

fix process.title

enter node's context before callback

chain PromiseRejectCallback

+++ deps/uv/src/unix/getaddrinfo.c
@@ -88,7 +88,7 @@ int uv__getaddrinfo_translate_error(int sys_err) {
   case EAI_SYSTEM: return -errno;
 #endif
   }
-  assert(!"unknown EAI_* error code");
+  assert(!(int)("unknown EAI_* error code"));
   abort();
   return 0;  /* Pacify compiler. */
 }
